requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
lxml==6.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyyaml>=6.0.1
pydantic-settings>=2.2
requests==2.32.4
soupsieve==2.7
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
